requirements.txt
blinker==1.9.0
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.19.1
Flask==3.1.2
fonttools==4.60.1
fsspec==2025.9.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
setuptools==70.2.0
six==1.17.0
sympy==1.14.0
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
triton==3.3.1
typing_extensions==4.15.0
Werkzeug==3.1.3
